backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=10.0.0

# Дополнительные зависимости для диагностики и разработки
requests>=2.31.0
httpx>=0.25.0

# Зависимости для Telegram бота
python-telegram-bot>=20.7
aiohttp>=3.9.0